refactor App, List, ListElement, change Component Composition
