+++ src/components/ErrorComp.jsx
@@ -1,11 +1,9 @@
-import './List.css';
-
 export default function ErrorComp() {
   return (
     <>
       <h1>Ups...</h1>
       <h2>We couldn´t get any wather-data, here are some ideas, what you can do:</h2>
-      <ul className="activityList">
+      <ul style={{ listStyleType: 'none' }}>
         <li>Read a book...</li>
         <li>Watch Tv...</li>
         <li>Surf on Reddit...</li>
