+++ src/components/ListElement.jsx
@@ -1,7 +1,12 @@
-import "./ListElement.css"
+import './ListElement.css';
 
-export default function ListElement ({activity}) {
-    return (
-        <li className="ListElement" >{activity.name}<button className="ListElement__button">x</button></li>
-    )
+export default function ListElement({ activity, onDelete }) {
+  return (
+    <li className="ListElement">
+      {activity.name}
+      <button onClick={() => onDelete(activity.id)} className="ListElement__button">
+        x
+      </button>
+    </li>
+  );
 }
